+++ src/Mods/AstralArena/ScriptExtender/Lua/AstralArena/Shared/Constants.lua
@@ -1,6 +1,6 @@
 return {
     ModuleUUID = "3b7d59a3-a846-4ecd-9c5d-a7bfe3f1b84f",
-    DisplayVersion = "0.2.0-alpha.1",
+    DisplayVersion = "0.2.1-alpha.1",
     SchemaVersion = 1,
     EntrantCount = 8,
     MaxPartySize = 4,